requirements.txt
fastapi
uvicorn
torch
torchvision
transformers
pillow
